added getHeight function to DRY
