+++ src/zedwallet++/CommandImplementations.cpp
@@ -265,8 +265,11 @@ void status(const std::shared_ptr<WalletBackend> walletBackend)
 }
 
 void reset(const std::shared_ptr<WalletBackend> walletBackend)
-{
-    const uint64_t scanHeight = ZedUtilities::getScanHeight();
+{   
+
+    const std::string msg = "What height would you like to begin scanning your wallet from?\n\nThis can greatly speed up the initial wallet scanning process.\n\nIf you do not know the exact height, err on the side of caution so transactions do not get missed.\n\nHit enter for the sub-optimal default of zero: ";
+
+    const uint64_t scanHeight = getHeight(msg);
 
     std::cout << std::endl
               << InformationMsg("This process may take some time to complete.") << std::endl
@@ -297,7 +300,13 @@ void reset(const std::shared_ptr<WalletBackend> walletBackend)
 
 void rewind(const std::shared_ptr<WalletBackend> walletBackend)
 {
-    const uint64_t scanHeight = ZedUtilities::getRewindToHeight(walletBackend);
+
+    const WalletTypes::WalletStatus status = walletBackend->getStatus();
+    const uint64_t defaultRewindHeight = status.walletBlockCount < 1000 ? 0 : status.walletBlockCount - 1000;
+
+    const std::string msg = "What block height do you want to rewind your wallet to?\n\nAll blocks after this height will be rescanned, use this command if you suspect a transaction has been missed by the sync process.\n\nHit enter for the default of " + std::to_string(defaultRewindHeight) + " (1000 blocks ago): ";
+
+    const uint64_t scanHeight = getHeight(msg);
 
     std::cout << std::endl
               << InformationMsg("This process may take some time to complete.") << std::endl
@@ -328,8 +337,27 @@ void rewind(const std::shared_ptr<WalletBackend> walletBackend)
 
 void scanRange(const std::shared_ptr<WalletBackend> walletBackend)
 {
-    const auto [startHeight, endHeight] = ZedUtilities::getScanRange();
 
+    const std::string startHeightMsg = "What height would you like to begin scanning your wallet from?\n\nThis can greatly speed up the initial wallet scanning process.\n\nIf you do not know the exact height, err on the side of caution so transactions do not get missed.\n\nHit enter for the sub-optimal default of zero:" ;
+    const uint64_t startHeight = getHeight(startHeightMsg);
+
+
+    std::string defaultEndHeight;
+    defaultEndHeight = std::to_string(startHeight + 1000);
+
+    const std::string endHeightMsg = "What height would you like to end scanning your wallet from?\n\nHit enter for the default of " + defaultEndHeight + ": ";
+
+    uint64_t endHeight = getHeight(endHeightMsg);
+
+    while(startHeight > endHeight)
+    {
+        std::cout << WarningMsg("The end block height should be greater than the starting height you provided(") 
+        << WarningMsg(std::to_string(startHeight)) 
+        << WarningMsg(")");
+        endHeight = getHeight(endHeightMsg);
+        
+    }
+   
     std::cout << std::endl
               << InformationMsg("This process may take some time to complete.") << std::endl
               << InformationMsg("You can't make any transactions during the ") << InformationMsg("process.")
@@ -343,14 +371,11 @@ void scanRange(const std::shared_ptr<WalletBackend> walletBackend)
 
     std::cout << InformationMsg("Syncing wallet...") << std::endl;
 
-    const uint64_t timestamp = 0;
-
     /* Don't want to queue up transaction events, since sync wallet will print
        them out */
     walletBackend->m_eventHandler->onTransaction.pause();
 
-    // walletBackend->reset(startHeight, timestamp);
-    walletBackend->scanRange(startHeight, endHeight, timestamp);
+    walletBackend->scanRange(startHeight, endHeight);
 
     syncWallet(walletBackend);
