+++ src/zedwallet++/GetInput.h
@@ -22,4 +22,6 @@ template<typename T> std::string getInput(const std::vector<T> &availableCommand
 
 std::tuple<std::string, uint16_t, bool> getDaemonAddress();
 
+uint64_t getHeight(const std::string msg);
+
 std::string getHash(const std::string msg, const bool cancelAllowed);
